Fix TypeError
I was getting "TypeError: Cannot call a class as a function" in the console in production builds on Vercel. I removed a bunch of code and discovered that the Temporal polyfill was the root cause of the issue. This commit removes Temporal and uses Luxon. The errors cleared as a result.