HT-3 Refactor Tooltip and XAxis
This commit improves the look and feel of the Tooltip, and also changes how the X Axis custom ticks are formatted. This allows us to keep all the dates in the data object instead of formatting the data object to fit how we wanted things to look in the X axis.